fix: aux functions returning int

+++ srcs/linear_algebra/matrices/matrix_utils.c
@@ -6,13 +6,13 @@
 /*   By: fnacarel <fnacarel@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/28 13:57:03 by fnacarel          #+#    #+#             */
-/*   Updated: 2023/08/02 18:31:34 by revieira         ###   ########.fr       */
+/*   Updated: 2023/08/03 14:56:09 by revieira         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../../includes/minirt.h"
 
-int	minor(t_matrix a, int row, int col)
+double	minor(t_matrix a, int row, int col)
 {
 	t_matrix	sub;
 
@@ -20,7 +20,7 @@ int	minor(t_matrix a, int row, int col)
 	return (sub.matr[0][0] * sub.matr[1][1] - sub.matr[0][1] * sub.matr[1][0]);
 }
 
-int	cofactor(t_matrix a, int row, int col)
+double	cofactor(t_matrix a, int row, int col)
 {
 	if (a.rows == 4 && a.cols == 4)
 	{
@@ -44,7 +44,7 @@ int	is_invertible(t_matrix a)
 	return (0);
 }
 
-int	determinant(t_matrix a)
+double	determinant(t_matrix a)
 {
 	int	i;
 	int	cof;
